refactor: Rename `lib/plugins/aws/package/lib/generateCoreTemplate.js`
