+++ lib/plugins/aws/package/index.js
@@ -4,7 +4,7 @@ const path = require('path');
 const mergeCustomProviderResources = require('./lib/mergeCustomProviderResources');
 const stripNullPropsFromTemplateResources = require('./lib/stripNullPropsFromTemplateResources');
 const generateArtifactDirectoryName = require('./lib/generate-artifact-directory-name');
-const generateCoreTemplate = require('./lib/generateCoreTemplate');
+const generateCoreTemplate = require('./lib/generate-core-template');
 const saveServiceState = require('./lib/saveServiceState');
 const saveCompiledTemplate = require('./lib/saveCompiledTemplate');
 const mergeIamTemplates = require('./lib/mergeIamTemplates');
