+++ lib/plugins/aws/package/lib/generate-core-template.js
