+++ test/unit/lib/plugins/aws/package/lib/generate-core-template.test.js
